requirements.txt
numpy>=1.23.0
pandas>=1.5.0
matplotlib>=3.6.0
scipy>=1.9.0
tensorflow>=2.12.0
lime>=0.2.0
scikit-image>=0.19.0
scikit-learn>=1.1.0
librosa>=0.10.0
tqdm>=4.64.0
umap-learn>=0.5.3
